+++ src/nistchempy/__init__.py
@@ -5,17 +5,21 @@ WebBook pages for research workflows. It is not affiliated with,
 maintained by, or endorsed by NIST.
 '''
 
-__version__ = '1.0.6'
-__updated__ = 'May 16, 2026'
+__version__ = '1.1.0.dev0'
+__updated__ = 'May 19, 2026'
 __license__ = 'MIT'
 
 
 from nistchempy.compound_list import get_all_data
+from nistchempy.index import WebBookIndex, get_local_index
 from nistchempy.compound import get_compound
 from nistchempy.search import run_search, run_structural_search
 from nistchempy.search import NistSearchParameters
 from nistchempy.search import get_search_parameters, print_search_parameters
 from nistchempy.requests import RequestConfig
+from nistchempy.exceptions import NistChemPyError
+from nistchempy.exceptions import NistChemPyIndexError
+from nistchempy.exceptions import NistChemPyIndexNotFoundError
 from nistchempy.utils import get_crawl_delay